% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_box.R
\name{.each_player_box}
\alias{.each_player_box}
\title{Parse player box score data for each game list}
\usage{
.each_player_box(resp)
}
\arguments{
\item{resp}{list element from scraped json data}
}
\description{
Internal function for parse_player_box
}
