% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_box_scores.R
\name{.each_team_box}
\alias{.each_team_box}
\title{Parse team box score data for each game list}
\usage{
.each_team_box(resp)
}
\arguments{
\item{resp}{list element from scraped json data}
}
\description{
Internal function for parse_team_box
}
