% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_shots.R
\name{.parse_each_match_shot}
\alias{.parse_each_match_shot}
\title{Parse shooting location data for single game}
\usage{
.parse_each_match_shot(resp)
}
\arguments{
\item{resp}{list element from scraped json data}
}
\description{
Internal function for parse_match_shot
}
