% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp.R
\name{each_pbp}
\alias{each_pbp}
\title{Parse pbp data for each game list}
\usage{
each_pbp(resp)
}
\arguments{
\item{resp}{list element from scraped json data}
}
\description{
This functions parses each game list and returns a 
data frame of pbp data
}
