% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_games_lists.R
\name{get_each_match}
\alias{get_each_match}
\title{Scrape each game json for season 25-26}
\usage{
get_each_match(match_id)
}
\arguments{
\item{match_id}{match id which can only be found in the html elements}
}
\description{
This functions scrapes each json file for each match number and returns
a data frame of required data, including nested data frames
}
