% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_games_lists.R
\name{get_games}
\alias{get_games}
\title{Scrape each game json}
\usage{
get_games(match_id)
}
\arguments{
\item{match_id}{match id can be found in each game url on nbl.com.au}
}
\description{
This functions scrapes each json file for each match number and returns
a list of data points
}
