% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{get_matches}
\alias{get_matches}
\title{Scrape matches for single season}
\usage{
get_matches(season_id)
}
\arguments{
\item{season_id}{can be the competitionId from in get_seasons()}
}
\description{
This function is a scraper for all match data for
a single season
}
