% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_shots.R
\name{parse_match_shot}
\alias{parse_match_shot}
\title{Parse shooting location data for game lists}
\usage{
parse_match_shot(resp)
}
\arguments{
\item{resp}{list element(s) from scraped json data}
}
\description{
This functions parses all game lists and returns a 
data frame of player shooting coordinates and types
}
