% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp.R
\name{parse_pbp}
\alias{parse_pbp}
\title{Parse pbp data for game lists}
\usage{
parse_pbp(resp)
}
\arguments{
\item{resp}{list element(s) from scraped json data}
}
\description{
This functions parses all game list and returns a 
data frame of pbp data
}
