% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_box.R
\name{parse_player_box}
\alias{parse_player_box}
\title{Parse player box score data for each game lists}
\usage{
parse_player_box(resp)
}
\arguments{
\item{resp}{list element(s) from scraped json data}
}
\description{
This functions parses all game lists and returns a 
data frame of player box score data
}
