% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_players}
\alias{parse_players}
\title{Parse scraped game list for player names}
\usage{
parse_players(resp)
}
\arguments{
\item{resp}{list element from scraped json data}
}
\description{
This functions parses scraped NBL match json for player identifying
details including, name, surname, shirt number, etc
}
