% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_box_scores.R
\name{parse_team_box}
\alias{parse_team_box}
\title{Parse team box score data for game lists}
\usage{
parse_team_box(resp)
}
\arguments{
\item{resp}{list element(s) from scraped json data}
}
\description{
This functions parses all game list and returns a 
data frame of team box score data
}
