% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_nblr}
\alias{save_nblr}
\title{Save files to nblr_data release}
\usage{
save_nblr(df, file_name, release_tag)
}
\arguments{
\item{df}{data_frame to save}

\item{file_name}{file_name to upload as, without the file extension}

\item{release_tag}{name of release to upload to}
}
\description{
This functions attaches nblr_data attributes like type and timestamp, saves
data to a temporary directory in rds format and then uploads to nblr_data
repository for a specified release tag.
}
